gormfunc->gorm name refactor
