+++ gorm/search.go
@@ -1,4 +1,4 @@
-package gormfunc
+package gorm
 
 import (
 	"gorm.io/gorm"
